Fix SLN
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@120 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/nanohttp.cpp
@@ -85,7 +85,7 @@ HTTP *HTTP::RequestURL(std::string url, IPAddress outgoing)
 	return new HTTP(host, port, path, outgoing);
 }
 
-void HTTP::Tick(int timeout)
+void HTTP::Tick(double timeout)
 {
 	this->client->Tick(timeout);
